backend/requirements/base.txt
Django==6.0.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
Pillow==11.0.0
psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0
stripe==11.3.0
python-dotenv==1.0.1
whitenoise==6.8.2
gunicorn==23.0.0
dj-database-url==2.3.0
